requirements.txt
nose
mock
ldap3
pymongo
py-bcrypt
flask
python-consul
biomaj_core
